+++ src/main/kotlin/com/example/olx/posts/repository/PostRepository.kt
@@ -4,4 +4,4 @@ import com.example.olx.posts.entity.Post
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor
 
-interface PostRepository : JpaRepository<Post, Int>, JpaSpecificationExecutor<Post> {}
+interface PostRepository : JpaRepository<Post, Long>, JpaSpecificationExecutor<Post> {}
